Feat: 상태 데코레이터, 선택 데코레이터 테스트
setDate == selDate 일 때 데코레이터 제대로 작동 x
