+++ app/src/main/res/drawable/cal_state3_select.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item>
+        <shape android:shape="rectangle" >
+            <solid android:color="@color/state3" />
+        </shape>
+    </item>
+
+    <item>
+        <shape android:shape="oval" >
+            <solid android:color="@color/black" />
+            <stroke
+                android:width="30dp"
+                android:color="@android:color/transparent" />
+        </shape>
+    </item>
+
+</layer-list>
