+++ app/src/main/res/drawable/decorator_state5_select.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector android:exitFadeDuration="100"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:state_checked="true"
+        android:drawable="@drawable/cal_state5_select" />
+    <item
+        android:state_pressed="true"
+        android:drawable="@drawable/cal_state5_select" />
+    <item android:drawable="@drawable/decorator_state5" />
+</selector>
